refactor: 🧩 Change fetcherUtils to fetcherSenderUtils
