+++ src/components/media/Friends.jsx
@@ -4,7 +4,7 @@ import Score from "./Score";
 import style from "./Friends.module.scss";
 import { A, useParams } from "@solidjs/router";
 import { useAuthentication, useMediaInfo } from "../../context/providers";
-import { apiRequestManager, fetcherUtils } from "../../utils/utils";
+import { apiRequestManager, fetcherSenderUtils, fetcherUtils } from "../../utils/utils";
 import { fetchers, fetcherSenders } from "../../collections/collections.js";
 
 function Friends() {
@@ -12,8 +12,8 @@ function Friends() {
   const { accessToken, authUserData } = useAuthentication();
 
   const { anilistData } = useMediaInfo();
-  const friendScoreFetcher = fetcherUtils.createFetcher(fetchers.anilist.getFrendScoresFromMedia, accessToken, () => params.id, { page: 1, perPage: 8 });
-  const cacheType = fetcherUtils.dynamicCacheType({
+  const friendScoreFetcher = fetcherSenderUtils.createFetcher(fetchers.anilist.getFrendScoresFromMedia, accessToken, () => params.id, { page: 1, perPage: 8 });
+  const cacheType = fetcherSenderUtils.dynamicCacheType({
     "default": () => apiRequestManager.anilist.inFiveSeconds() > 1,
     "only-if-cached": () => apiRequestManager.anilist.inFiveSeconds() > 2
   });
