+++ src/utils/utils.js
@@ -1,5 +1,6 @@
 export * as arrayUtils from "./arrays.js";
 export * as functionUtils from "./functionUtils.js";
+export * as fetcherSenderUtils from "./fetcherSenderUtils.js";
 export * as fetcherUtils from "./Fetcher/fetcherUtils.js";
 export * as scheduleUtils from "./scheduled.js";
 export * as dateUtils from "./dateUtils.js";
